telegram_bot/requirements.txt
aiogram==3.15.0
python-dotenv==1.0.1
gspread==6.1.4
oauth2client==4.1.3
asyncpg==0.29.0
redis==5.2.1
aioredis==2.0.1
sqlalchemy==2.0.36
